added better test coverage

+++ services/web/server/src/simcore_service_webserver/users/settings.py
@@ -8,7 +8,7 @@ from .._constants import APP_SETTINGS_KEY
 
 class UsersSettings(BaseCustomSettings, MixinServiceSettings):
     USERS_FRONTEND_PREFERENCES_DEFAULTS_OVERWRITES: dict[
-        str, float | int | str | list | dict | None
+        str, int | float | str | list | dict | None
     ] = Field(
         default_factory=dict,
         description="key: name of the FrontendUserPreference, value: new default",
